Finished PetProvider query() method
Catalog Activity no longer touches the db directly. Instead it calls a
content resolver with the content uri and other paramters as input,
which calls the correct provider. The PetProvider now includes a URI
Matcher which takes the query parameters and matches/validates before
allowing the PetProvider to complete the query.